call back qrcode

+++ Famelicus/app/src/main/java/br/ufrj/cos/famelicus/Proxy.java
@@ -26,8 +26,8 @@ public class Proxy {
         System.out.println(toServer);
     };
 
-    public float pedirVersaoBD(){
-        float version = 0;
+    public int pedirVersaoBD(){
+        int version = 0;
         return version;
     };
